backend/src/main/java/com/woowacourse/f12/application/ReviewService.java
package com.woowacourse.f12.application;

import com.woowacourse.f12.domain.Review;
import com.woowacourse.f12.domain.ReviewRepository;
import com.woowacourse.f12.dto.request.ReviewRequest;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional(readOnly = true)
public class ReviewService {

    private final ReviewRepository reviewRepository;

    public ReviewService(final ReviewRepository reviewRepository) {
        this.reviewRepository = reviewRepository;
    }

    @Transactional
    public Long save(final Long productId, final ReviewRequest reviewRequest) {
        final Review review = reviewRequest.toReview(productId);
        return reviewRepository.save(review)
                .getId();
    }
}
